+++ consensus/bor/finality/whitelist_helpers.go
@@ -1,12 +1,12 @@
-package borfinality
+package finality
 
 import (
 	"context"
 	"errors"
 
 	"github.com/ledgerwatch/erigon-lib/common"
+	"github.com/ledgerwatch/erigon/consensus/bor/finality/whitelist"
 	"github.com/ledgerwatch/erigon/consensus/bor/heimdall"
-	"github.com/ledgerwatch/erigon/eth/borfinality/whitelist"
 	"github.com/ledgerwatch/log/v3"
 )
